# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: SST Auditor Agent Pro — Backend API
  version: "0.1.0"
  description: >
    Especificación de referencia de los endpoints expuestos por el backend.
    Se actualiza en cada ticket que agregue o modifique endpoints.
paths:
...
  /api/v1/calificaciones-estandar/{calificacion_id}/evidencias:
    post:
      summary: Registra metadatos de un documento de soporte (sin el binario)
      tags: [Evidencias]
      security: [{bearerAuth: []}]
      parameters:
        - name: calificacion_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudRegistrarEvidencia"
      responses:
        "201":
          description: Metadatos persistidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaEvidencia"
        "403":
          description: ACCESO_DENEGADO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "404":
          description: CALIFICACION_NO_ENCONTRADA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "422":
          description: EVIDENCIA_INVALIDA / ERROR_VALIDACION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        default:
          $ref: "#/components/responses/ErrorPorDefecto"
    get:
      summary: Lista evidencias activas de una calificación
      tags: [Evidencias]
      security: [{bearerAuth: []}]
      parameters:
        - name: calificacion_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Solo filas con activo=true
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaEvidencia"
        "404":
          description: CALIFICACION_NO_ENCONTRADA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        default:
          $ref: "#/components/responses/ErrorPorDefecto"
  /api/v1/evidencias/{id}:
    delete:
      summary: Da de baja lógica una evidencia
      tags: [Evidencias]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "204":
          description: Marcada inactiva; la fila se conserva
        "403":
          description: ACCESO_DENEGADO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "404":
          description: EVIDENCIA_NO_ENCONTRADA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "409":
          description: EVIDENCIA_YA_INACTIVA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        default:
          $ref: "#/components/responses/ErrorPorDefecto"
  /api/v1/evidencias/{id}/enlace-descarga:
    post:
      summary: Emite un enlace temporal de descarga
      tags: [Evidencias]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: URL relativa con JWT tipo=descarga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaEnlaceDescarga"
        "403":
          description: ACCESO_DENEGADO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "404":
          description: EVIDENCIA_NO_ENCONTRADA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        default:
          $ref: "#/components/responses/ErrorPorDefecto"
  /api/v1/descargas/evidencias:
    get:
      summary: Canjea un enlace de descarga
      tags: [Evidencias]
      parameters:
        - name: token
          in: query
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Binario, si existe bajo ALMACENAMIENTO_LOCAL_RAIZ
        "401":
          description: TOKEN_INVALIDO o TOKEN_EXPIRADO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "403":
          description: ACCESO_DENEGADO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "404":
          description: EVIDENCIA_NO_ENCONTRADA o ARCHIVO_NO_DISPONIBLE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        default:
          $ref: "#/components/responses/ErrorPorDefecto"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de acceso (claim `tipo=acceso`). No aceptar tokens de refresco.
  schemas:
...
    SolicitudRegistrarEvidencia:
      type: object
      additionalProperties: false
      required: [nombre_archivo, tipo_mime, tamano_bytes, ruta_almacenamiento]
      properties:
        nombre_archivo: {type: string, maxLength: 255}
        tipo_mime:
          type: string
          enum: [application/pdf, image/jpeg, image/png]
        tamano_bytes: {type: integer, minimum: 1, maximum: 10485760}
        ruta_almacenamiento: {type: string, maxLength: 500}
    RespuestaEvidencia:
      type: object
      required:
        - id
        - calificacion_estandar_id
        - usuario_id
        - nombre_archivo
        - tipo_mime
        - tamano_bytes
        - ruta_almacenamiento
        - fecha_carga
        - activo
      properties:
        id: {type: string, format: uuid}
        calificacion_estandar_id: {type: string, format: uuid}
        usuario_id: {type: string, format: uuid}
        nombre_archivo: {type: string}
        tipo_mime: {type: string}
        tamano_bytes: {type: integer}
        ruta_almacenamiento: {type: string}
        fecha_carga: {type: string, format: date-time}
        activo: {type: boolean}
    RespuestaEnlaceDescarga:
      type: object
      required: [url, expira_en_segundos]
      properties:
        url: {type: string}
        expira_en_segundos: {type: integer, maximum: 900}
...
    RespuestaError:
      type: object
      description: >
        Contrato único de toda respuesta de error del API (4xx y 5xx).
        Todas las rutas presentes y futuras responden errores con este esquema.
      required: [exito, codigo, mensaje]
      properties:
        exito:
          type: boolean
          description: Siempre `false` en respuestas de error.
          example: false
        codigo:
          type: string
          description: >
            Código estable en mayúsculas: ERROR_DOMINIO, ERROR_VALIDACION,
            ERROR_HTTP, ERROR_INTERNO o el código específico de una subclase
            de DomainException.
          example: ERROR_VALIDACION
        mensaje:
          type: string
          description: Descripción legible del error, en español.
          example: La petición contiene datos inválidos.
        detalle:
          description: >
            Información adicional (p. ej. errores de campo de Pydantic).
            `null` cuando no aplica.
          oneOf:
            - type: array
              items:
                type: object
            - type: object
            - type: "null"
  responses:
    ErrorPorDefecto:
      description: Respuesta de error con el esquema único `RespuestaError`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespuestaError"
          examples:
            errorValidacion:
              summary: 422 — error de validación
              value:
                exito: false
                codigo: ERROR_VALIDACION
                mensaje: La petición contiene datos inválidos.
                detalle:
                  - loc: [body, name]
                    msg: Field required
                    type: missing
            errorInterno:
              summary: 500 — error no controlado
              value:
                exito: false
                codigo: ERROR_INTERNO
                mensaje: Ha ocurrido un error interno. Contacte al administrador.
                detalle: null